+++ buildSrc/src/main/kotlin/BuildSupport.gradle.kts
@@ -34,6 +34,8 @@ repositories {
   }
 }
 
+testTasks()
+
 tasks.withType<Jar>().configureEach {
   manifest {
     attributes["Implementation-Title"] = "iceberg-catalog-migrator"
@@ -61,3 +63,45 @@ plugins.withType<JavaPlugin>().configureEach {
     modularity.inferModulePath.set(true)
   }
 }
+
+fun Project.testTasks() {
+  if (projectDir.resolve("src/test").exists()) {
+    tasks.withType<Test>().configureEach {
+      useJUnitPlatform {}
+      val testJvmArgs: String? by project
+      if (testJvmArgs != null) {
+        jvmArgs((testJvmArgs as String).split(" "))
+      }
+
+      systemProperty("file.encoding", "UTF-8")
+      systemProperty("user.language", "en")
+      systemProperty("user.country", "US")
+      systemProperty("user.variant", "")
+      systemProperty("test.log.level", testLogLevel())
+      environment("TESTCONTAINERS_REUSE_ENABLE", "true")
+      filter {
+        isFailOnNoMatchingTests = false
+        when (name) {
+          "test" -> {
+            includeTestsMatching("*Test")
+            includeTestsMatching("Test*")
+            excludeTestsMatching("Abstract*")
+            excludeTestsMatching("IT*")
+          }
+          "intTest" -> includeTestsMatching("IT*")
+        }
+      }
+      if (name != "test") {
+        mustRunAfter(tasks.named<Test>("test"))
+      }
+    }
+    val intTest =
+      tasks.register<Test>("intTest") {
+        group = "verification"
+        description = "Runs the integration tests."
+      }
+    tasks.named("check") { dependsOn(intTest) }
+  }
+}
+
+fun testLogLevel() = System.getProperty("test.log.level", "WARN")
