+++ cli/src/exec/exec-preamble.sh
