+++ cli/src/main/resources/logback.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!--
+
+    Copyright (C) 2023 Dremio
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+-->
+<configuration debug="true">
+  <property name="catalog.migration.log.dir" value="${catalog.migration.log.dir:-logs}"/>
+  <appender name="FILE" class="ch.qos.logback.core.FileAppender">
+    <file>${catalog.migration.log.dir}/catalog_migration.log</file>
+    <append>true</append>
+    <encoder>
+      <pattern>%date{ISO8601} [%thread] %-5level %logger{36} - %msg%n</pattern>
+    </encoder>
+  </appender>
+
+  <logger name="org.projectnessie.tools.catalog.migration"/>
+  <root>
+    <level value="INFO"/>
+    <appender-ref ref="FILE"/>
+  </root>
+</configuration>
