+++ cli/src/main/resources/org/projectnessie/tools/catalog/migration/cli/version.properties
